+++ activejob/lib/active_job/arguments.rb
@@ -98,7 +98,7 @@ module ActiveJob
               Primitive serialization of BigDecimal job arguments is deprecated as it may serialize via .to_s using certain queue adapters.
               Enable config.active_job.use_big_decimal_serializer to use BigDecimalSerializer instead, which will be mandatory in Rails 7.2.
 
-              Note that if you application has multiple replicas, you should only enable this setting after successfully deploying your app to Rails 7.1 first.
+              Note that if your application has multiple replicas, you should only enable this setting after successfully deploying your app to Rails 7.1 first.
               This will ensure that during your deployment all replicas are capable of deserializing arguments serialized with BigDecimalSerializer.
             MSG
             return argument