+++ activemodel/lib/active_model/attribute_registration.rb
@@ -10,17 +10,28 @@ module ActiveModel
 
     module ClassMethods # :nodoc:
       def attribute(name, type = nil, default: (no_default = true), **options)
+        name = resolve_attribute_name(name)
         type = resolve_type_name(type, **options) if type.is_a?(Symbol)
+        type = hook_attribute_type(name, type) if type
 
-        pending = pending_attribute(name)
-        pending.type = type if type
-        pending.default = default unless no_default
+        pending_attribute_modifications << PendingType.new(name, type) if type || no_default
+        pending_attribute_modifications << PendingDefault.new(name, default) unless no_default
+
+        reset_default_attributes
+      end
+
+      def decorate_attributes(names = nil, &decorator) # :nodoc:
+        names = names&.map { |name| resolve_attribute_name(name) }
+
+        pending_attribute_modifications << PendingDecorator.new(names, decorator)
 
         reset_default_attributes
       end
 
       def _default_attributes # :nodoc:
-        @default_attributes ||= build_default_attributes
+        @default_attributes ||= AttributeSet.new({}).tap do |attribute_set|
+          apply_pending_attribute_modifications(attribute_set)
+        end
       end
 
       def attribute_types # :nodoc:
@@ -30,39 +41,51 @@ module ActiveModel
       end
 
       private
-        class PendingAttribute # :nodoc:
-          attr_accessor :type, :default
-
-          def apply_to(attribute)
-            attribute = attribute.with_type(type || attribute.type)
-            attribute = attribute.with_user_default(default) if defined?(@default)
-            attribute
+        PendingType = Struct.new(:name, :type) do # :nodoc:
+          def apply_to(attribute_set)
+            attribute = attribute_set[name]
+            attribute_set[name] = attribute.with_type(type || attribute.type)
           end
         end
 
-        def pending_attribute(name)
-          @pending_attributes ||= {}
-          @pending_attributes[resolve_attribute_name(name)] ||= PendingAttribute.new
+        PendingDefault = Struct.new(:name, :default) do # :nodoc:
+          def apply_to(attribute_set)
+            attribute_set[name] = attribute_set[name].with_user_default(default)
+          end
         end
 
-        def apply_pending_attributes(attribute_set)
-          superclass.send(__method__, attribute_set) if superclass.respond_to?(__method__, true)
+        PendingDecorator = Struct.new(:names, :decorator) do # :nodoc:
+          def apply_to(attribute_set)
+            (names || attribute_set.keys).each do |name|
+              attribute = attribute_set[name]
+              type = decorator.call(name, attribute.type)
+              attribute_set[name] = attribute.with_type(type) if type
+            end
+          end
+        end
 
-          defined?(@pending_attributes) && @pending_attributes.each do |name, pending|
-            attribute_set[name] = pending.apply_to(attribute_set[name])
+        def pending_attribute_modifications
+          @pending_attribute_modifications ||= []
+        end
+
+        def apply_pending_attribute_modifications(attribute_set)
+          if superclass.respond_to?(:apply_pending_attribute_modifications, true)
+            superclass.send(:apply_pending_attribute_modifications, attribute_set)
           end
 
-          attribute_set
-        end
-
-        def build_default_attributes
-          apply_pending_attributes(AttributeSet.new({}))
+          pending_attribute_modifications.each do |modification|
+            modification.apply_to(attribute_set)
+          end
         end
 
         def reset_default_attributes
+          reset_default_attributes!
+          subclasses.each { |subclass| subclass.send(:reset_default_attributes) }
+        end
+
+        def reset_default_attributes!
           @default_attributes = nil
           @attribute_types = nil
-          subclasses.each { |subclass| subclass.send(__method__) }
         end
 
         def resolve_attribute_name(name)
@@ -71,6 +94,13 @@ module ActiveModel
 
         def resolve_type_name(name, **options)
           Type.lookup(name, **options)
+        end
+
+        # Hook for other modules to override. The attribute type is passed
+        # through this method immediately after it is resolved, before any type
+        # decorations are applied.
+        def hook_attribute_type(attribute, type)
+          type
         end
     end
   end