+++ guides/bug_report_templates/action_controller_main.rb
@@ -8,7 +8,6 @@ gemfile(true) do
   git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 
   gem "rails", github: "rails/rails", branch: "main"
-  gem "rack", "~> 2.0"
 end
 
 require "action_controller/railtie"